v2/requirements.txt
pydantic>=2.0.0
pydantic-ai>=0.0.46
mcp>=1.6.0
openai>=1.3.0
python-dotenv>=1.0.0
rich>=13.0.0
anyio>=3.7.0
httpx>=0.24.0
logfire>=0.6.0
